hclus {radiant.multivariate}	R Documentation

Hierarchical cluster analysis

Description

Hierarchical cluster analysis

Usage

hclus(dataset, vars, labels = "none", distance = "sq.euclidian",
  method = "ward.D", max_cases = 5000, data_filter = "")

Arguments

dataset
Dataset

vars
Vector of variables to include in the analysis

labels
A vector of labels for the leaves of the tree

distance
Distance

method
Method

max_cases
Maximum number of cases allowed (default is 1000). Set to avoid long-running analysis in the radiant web-interface

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

Details

See https://radiant-rstats.github.io/docs/multivariate/hclus.html for an example in Radiant

Value

A list of all variables used in hclus as an object of class hclus

See Also

summary.hclus to summarize results

plot.hclus to plot results

Examples

hclus(shopping, vars = "v1:v6") %>% str()

[Package radiant.multivariate version 0.9.9 Index]