kclus {radiant.multivariate}	R Documentation

K-clustering

Description

K-clustering

Usage

kclus(dataset, vars, fun = "mean", hc_init = TRUE,
  distance = "sq.euclidian", method = "ward.D", seed = 1234,
  nr_clus = 2, data_filter = "")

Arguments

dataset
Dataset

vars
Vector of variables to include in the analysis

fun
Function to use: "mean" or "median"

hc_init
Use centers from hclus as the starting point

distance
Distance for hclus

method
Method for hclus

seed
Random see to use for k-clustering if hc_init is FALSE

nr_clus
Number of clusters to extract

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

Details

See https://radiant-rstats.github.io/docs/multivariate/kclus.html for an example in Radiant

Value

A list of all variables used in kclus as an object of class kclus

See Also

summary.kclus to summarize results

plot.kclus to plot results

store.kclus to add cluster membership to the selected dataset

Examples

kclus(shopping, c("v1:v6"), nr_clus = 3) %>% str()

[Package radiant.multivariate version 0.9.9 Index]