plot.full_factor {radiant.multivariate}	R Documentation

Plot method for the full_factor function

Description

Plot method for the full_factor function

Usage

## S3 method for class 'full_factor'
plot(x, plots = "attr", shiny = FALSE,
  custom = FALSE, ...)

Arguments

x
Return value from full_factor

plots
Include attribute ("attr"), respondents ("resp") or both in the plot

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/multivariate/full_factor.html for an example in Radiant

See Also

full_factor to calculate results

plot.full_factor to plot results

Examples

result <- full_factor(shopping , "v1:v6", nr_fact = 2)
plot(result)

[Package radiant.multivariate version 0.9.9 Index]