plot.kclus {radiant.multivariate}	R Documentation

Plot method for kclus

Description

Plot method for kclus

Usage

## S3 method for class 'kclus'
plot(x, plots = "density", shiny = FALSE,
  custom = FALSE, ...)

Arguments

x
Return value from kclus

plots
One of "density", "bar", or "scatter")

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/multivariate/kclus.html for an example in Radiant

See Also

kclus to generate results

summary.kclus to summarize results

store.kclus to add cluster membership to the selected dataset

Examples

result <- kclus(shopping, vars = "v1:v6", nr_clus = 3)
plot(result)

[Package radiant.multivariate version 0.9.9 Index]