plot.mds {radiant.multivariate}	R Documentation

Plot method for the mds function

Description

Plot method for the mds function

Usage

## S3 method for class 'mds'
plot(x, rev_dim = NULL, fontsz = 5, shiny = FALSE,
  custom = FALSE, ...)

Arguments

x
Return value from mds

rev_dim
Flip the axes in plots

fontsz
Font size to use in plots

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/multivariate/mds.html for an example in Radiant

See Also

mds to calculate results

summary.mds to plot results

Examples

result <- mds(city, "from", "to", "distance")
plot(result, fontsz = 7)
plot(result, rev_dim = 1:2)

[Package radiant.multivariate version 0.9.9 Index]