predict.conjoint {radiant.multivariate}	R Documentation

Predict method for the conjoint function

Description

Predict method for the conjoint function

Usage

## S3 method for class 'conjoint'
predict(object, pred_data = NULL, pred_cmd = "",
  conf_lev = 0.95, se = FALSE, interval = "confidence", dec = 3,
  ...)

Arguments

object
Return value from conjoint

pred_data
Provide the dataframe to generate predictions. The dataset must contain all columns used in the estimation

pred_cmd
Command used to generate data for prediction

conf_lev
Confidence level used to estimate confidence intervals (.95 is the default)

se
Logical that indicates if prediction standard errors should be calculated (default = FALSE)

interval
Type of interval calculation ("confidence" or "prediction"). Set to "none" if se is FALSE

dec
Number of decimals to show

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/multivariate/conjoint.html for an example in Radiant

See Also

conjoint to generate the result

summary.conjoint to summarize results

plot.conjoint to plot results

Examples

result <- conjoint(mp3, rvar = "Rating", evar = "Memory:Shape")
predict(result, pred_data = mp3)

[Package radiant.multivariate version 0.9.9 Index]