predict_conjoint_by {radiant.multivariate}	R Documentation

Predict method for the conjoint function when a by variables is used

Description

Predict method for the conjoint function when a by variables is used

Usage

predict_conjoint_by(object, pfun, pred_data = NULL, pred_cmd = "",
  conf_lev = 0.95, se = FALSE, dec = 3, ...)

Arguments

object
Return value from conjoint

pfun
Function to use for prediction

pred_data
Name of the dataset to use for prediction

pred_cmd
Command used to generate data for prediction

conf_lev
Confidence level used to estimate confidence intervals (.95 is the default)

se
Logical that indicates if prediction standard errors should be calculated (default = FALSE)

dec
Number of decimals to show

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/multivariate/conjoint.html for an example in Radiant

See Also

conjoint to generate the result

summary.conjoint to summarize results

plot.conjoint to plot results

[Package radiant.multivariate version 0.9.9 Index]