store.kclus {radiant.multivariate}	R Documentation

Add a cluster membership variable to the active dataset

Description

Add a cluster membership variable to the active dataset

Usage

## S3 method for class 'kclus'
store(dataset, object, name = "", ...)

Arguments

dataset
Dataset to append to cluster membership variable to

object
Return value from kclus

name
Name of cluster membership variable

...
Additional arguments

Details

See https://radiant-rstats.github.io/docs/multivariate/kclus.html for an example in Radiant

See Also

kclus to generate results

summary.kclus to summarize results

plot.kclus to plot results

Examples

kclus(shopping, vars = "v1:v6", nr_clus = 3) %>%
  store(shopping, .) %>%
  head()

[Package radiant.multivariate version 0.9.9 Index]